solutions/iot/weather-stations/producer/requirements.txt
paho-mqtt
confluent_kafka
prometheus-kafka-metrics
prometheus-client
